requirements.txt
asgiref==3.5.0
autopep8==1.6.0
Django==4.0.3
django-ckeditor==6.2.0
django-js-asset==2.0.0
Pillow==9.0.1
pycodestyle==2.8.0
sqlparse==0.4.2
toml==0.10.2
tzdata==2021.5
